requirements.txt
torch==2.6.0+cu118
torchvision
pillow==11.0.0